+++ src/creatures/npcs/npc.cpp
@@ -201,7 +201,7 @@ void Npc::onThinkSound(uint32_t interval)
 		soundTicks = 0;
 
 		if (!npcType->info.soundVector.empty() && (npcType->info.soundChance >= static_cast<uint32_t>(uniform_random(1, 100)))) {
-			uint32_t index = uniform_random(0, npcType->info.soundVector.size() - 1);
+			uint32_t index = uniform_random(0, static_cast<int64_t>(npcType->info.soundVector.size() - 1));
 			g_game().sendSingleSoundEffect(this->getPosition(), npcType->info.soundVector[index], this);
 		}
 	}
@@ -411,7 +411,7 @@ void Npc::onThinkYell(uint32_t interval)
 		yellTicks = 0;
 
 		if (!npcType->info.voiceVector.empty() && (npcType->info.yellChance >= static_cast<uint32_t>(uniform_random(1, 100)))) {
-			uint32_t index = uniform_random(0, npcType->info.voiceVector.size() - 1);
+			uint32_t index = uniform_random(0, static_cast<int64_t>(npcType->info.voiceVector.size() - 1));
 			const voiceBlock_t& vb = npcType->info.voiceVector[index];
 
 			if (vb.yellText) {
