+++ src/utils/tools.h
@@ -54,8 +54,8 @@ constexpr bool hasBitSet(uint32_t flag, uint32_t flags) {
 }
 
 std::mt19937& getRandomGenerator();
-int32_t uniform_random(int32_t minNumber, int32_t maxNumber);
-int32_t normal_random(int32_t minNumber, int32_t maxNumber);
+int64_t uniform_random(int64_t minNumber, int64_t maxNumber);
+int64_t normal_random(int64_t minNumber, int64_t maxNumber);
 bool boolean_random(double probability = 0.5);
 
 Direction getDirection(const std::string& string);
